Midsem_project/requirements.txt
PyPDF2==3.0.1
nltk==3.8.1
spacy==3.7.4